onStop in Terminal component

+++ packages/terminal/main/TerminalChannel.js
@@ -5,11 +5,15 @@ class TerminalChannel extends IpcChannel {
   constructor (uid, cwd) {
     super('terminal', uid)
     this.pty = new Pty(this, cwd)
+    this.stopCommand = undefined
   }
 
   async run (command, config) {
     if (!command.trim()) {
       return
+    }
+    if (config.stop) {
+      this.stopCommand = config.stop.command
     }
     return await this.pty.run(command.trim(), config)
   }
@@ -21,7 +25,11 @@ class TerminalChannel extends IpcChannel {
   async kill () {
     return new Promise(resolve => {
       setTimeout(resolve, 2000)
-      this.pty.kill().then(resolve)
+      if (this.stopCommand) {
+        this.exec(this.stopCommand).then(resolve)
+      } else {
+        this.pty.kill().then(resolve)
+      }
     })
   }
 }
